backend/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydub
pyannote.audio
torchaudio
pdfplumber
python-docx
newspaper3k

